Change authentication middleware to only apply to endpoints that need it
