+++ server/routes/preferences-routes.js
@@ -1,10 +1,11 @@
 const preferencesController = require('../controller/preferences-controller');
+const { authenticate } = require('../utils/databaseUtils');
 const express = require('express');
 const router = express.Router();
 
 router.route('/')
-    .get(preferencesController.getPreferences)
-    .post(preferencesController.createPreferences)
-    .patch(preferencesController.changePreferences);
+    .get(authenticate, preferencesController.getPreferences)
+    .post(authenticate, preferencesController.createPreferences)
+    .patch(authenticate, preferencesController.changePreferences);
 
 module.exports = router;
